+++ src/scene.rs
@@ -1,6 +1,6 @@
 use bevy::{ecs::entity::hash_map::EntityHashMap, prelude::*, scene::DynamicEntity};
 
-use crate::{core::replication::replication_rules::ReplicationRules, Replicated};
+use crate::{Replicated, core::replication::replication_rules::ReplicationRules};
 
 /**
 Fills scene with all replicated entities and their components.