+++ CHANGELOG.md
@@ -7,11 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `ConditionerConfig` for `bevy_replicon_example_backend` to simulate various network conditions.
+
+### Changed
+
+- Don't insert `ClientVisibility` at all if `ServerPlugin::visibility_policy` is set to `VisibilityPolicy::All`. Previously all calls were just no-op.
+
 ## [0.31.1] - 2025-03-15
+
+### Changed
+
+- Rename `ClientId` into `NetworkId`.
+- Move `ConnectedClient::id` into a separate optional component for backends that doesn't provide persistent identifiers.
 
 ### Fixed
 
 - Bump the `bevy` dependency to 0.15.3 since we use some fields that were made public in a patch release.
+
+### Removed
+
+- `ConnectedClient::new`, you can now construct the struct directly.
 
 ## [0.31.0] - 2025-03-13